refactor: relocate internal utils to core folder
